+++ htdocs/SRSource.php
@@ -29,13 +29,6 @@ page_close();
 <link rel="stylesheet"
       href="css/table.css?version=<?= time() ?>"
 >
-
-<script src="js/react.development.js"
-        crossorigin
-></script>
-<script src="js/react-dom.development.js"
-        crossorigin
-></script>
 <script type="module"
-        src='components/CMPSRSource.js?version=<?= time() ?>'
+        src='components/dist/CMPSRSource.js?version=<?= time() ?>'
 ></script>